Worlds

+++ src/main/java/frc/robot/Subsystems/Shooter.java
@@ -56,15 +56,15 @@ public class Shooter {
     new Point2D.Double(83, 2600),
     new Point2D.Double(102, 2650),
     new Point2D.Double(120, 2800),
-    new Point2D.Double(138, 2900),
-    new Point2D.Double(156, 2975),
-    new Point2D.Double(174, 3195),
+    new Point2D.Double(138, 2950),
+    new Point2D.Double(156, 3100),
+    new Point2D.Double(174, 3220),
     new Point2D.Double(186, 3225),
     new Point2D.Double(210, 3400),
-    new Point2D.Double(225, 3475),
-    new Point2D.Double(240, 3610),
-    new Point2D.Double(264, 3625),
-    new Point2D.Double(276, 3715)
+    new Point2D.Double(225, 3600),
+    new Point2D.Double(240, 3650),
+    new Point2D.Double(264, 3900),
+    new Point2D.Double(276, 3975)
   };
   private LinearInterpolationTable m_rpmTable = new LinearInterpolationTable(rpmTable);
 
@@ -75,13 +75,13 @@ public class Shooter {
     new Point2D.Double(102, 11),
     new Point2D.Double(120, 18),
     new Point2D.Double(138, 21),
-    new Point2D.Double(156, 26),
-    new Point2D.Double(174, 28),
-    new Point2D.Double(186, 29),
-    new Point2D.Double(210, 30),
-    new Point2D.Double(225, 32),
-    new Point2D.Double(240, 32),
-    new Point2D.Double(264, 32),
+    new Point2D.Double(156, 24),
+    new Point2D.Double(174, 26),
+    new Point2D.Double(186, 28),
+    new Point2D.Double(210, 28),
+    new Point2D.Double(225, 28),
+    new Point2D.Double(240, 29),
+    new Point2D.Double(264, 30),
     new Point2D.Double(276, 32)
   };
   private LinearInterpolationTable m_hoodTable = new LinearInterpolationTable(hoodTable);
